Feat: 뷰어별 분류 박스 생성 다르게(#26)
